requirements.txt
Django==5.1.1
gunicorn==23.0.0
whitenoise==6.9.0
python-decouple==3.8
python-dotenv==1.0.1
django-jazzmin==3.0.1
django-allauth==65.7.0
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
Pillow==10.4.0
